requirements.txt
# Core Application & AI
streamlit
dspy
openai

# Database
psycopg[binary]

# Utilities
python-dotenv
faker 